refactor: use only sharding compiler
